+++ NEWS.md
@@ -9,6 +9,7 @@
 * added policy graph visualization with vizNetwork.
 * added plot_value_function.
 * added function sample_belief_space to sample from the belief space.
+* added function plot_belief_space.
 
 
 # pomdp 0.9.2 (12/06/2019)